+++ src/lib/terra-components.module.ts
@@ -18,6 +18,7 @@ import { ModalModule } from 'ngx-bootstrap/modal';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
 import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
@@ -39,6 +40,7 @@ import { MatInputModule } from '@angular/material/input';
         MatIconModule,
         MatCheckboxModule,
         MatFormFieldModule,
+        MatSelectModule,
         MatInputModule
     ],
     declarations: [...components, ...directives, ...pipes],